[feat] #52 Total List StateHolder 파일 분리
