+++ Techome/Techome/Views/TotalList/TotalListView.swift
@@ -48,46 +48,9 @@ struct TotalListLayoutValue {
     }
 }
 
-
-//카페인 + 부작용 리스트 병합에 사용
 struct TotalDataCell : Hashable {
     var dataType: String
     var dataIndex: Int
-}
-
-final class datas: ObservableObject { //observable 객체 생성
-    
-    @Published var dataSortedByDate = [ String : [TotalDataCell] ]()
-    @Published var offsetsArr = [ String : [CGFloat] ]()
-    @Published var datesArr = [String]()
-        
-    init(dataSortedByDate : [ String : [TotalDataCell]] ) {
-        self.dataSortedByDate = dataSortedByDate
-        
-        for key in self.dataSortedByDate.keys {
-            self.offsetsArr[key] = [CGFloat](repeating: .zero, count: self.dataSortedByDate[key]!.count)
-            self.datesArr.append(key)
-        }
-        
-        self.datesArr = datesArr.sorted(by: >)
-    }
-    
-    //선택된 cell 삭제
-    func deleteData(curDate : String, index : Int) {
-        self.dataSortedByDate[curDate]!.remove(at: index)
-        self.offsetsArr[curDate]!.remove(at:index)
-        
-        if (self.dataSortedByDate[curDate]!.count <= 0) {
-            self.datesArr.remove(at: self.datesArr.firstIndex(of: curDate)!)
-        }
-    }
-    
-    //모든 cell 원위치
-    func resetOffsets() {
-        for date in self.datesArr {
-            self.offsetsArr[date] = [CGFloat](repeating: .zero, count: dataSortedByDate[date]!.count)
-        }
-    }
 }
 
 // 전체 리스트
